Replace the Kite based implementation with a gobblin based one
This is done in order to meet the PDP-2 requirements which include:
- maintaining an identical time based path partitioning but with
a flexible prefix depending on the source info presence.
- adding avro header meta data fields with information about the
envelop's fields origin and a family_id field.

The original PNDA converter and writer are left unaltered as the new
implementation is being provided behind an experimental flag.
The converter and writer selection is done through the gobblin
configuration file.

PNDA-4464